spectral intensities to float

+++ BaseLibS/Ms/RawFile.cs
@@ -267,11 +267,11 @@ namespace BaseLibS.Ms{
 		}
 
 		protected internal abstract void GetSpectrum(int scanNumberMin, int scanNumberMax, int imsIndexMin,
-			int imsIndexMax, bool readCentroids, out double[] masses, out double[] intensities, double resolution,
+			int imsIndexMax, bool readCentroids, out double[] masses, out float[] intensities, double resolution,
 			double mzMin, double mzMax);
 
 		protected internal void GetSpectrum(int scanNumber, bool readCentroids, out double[] masses,
-			out double[] intensities){
+			out float[] intensities){
 			GetSpectrum(scanNumber, scanNumber, 0, 0, readCentroids, out masses, out intensities, 0, double.NaN,
 				double.NaN);
 		}
